% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/zzz.R
\name{to_kebab_case}
\alias{to_kebab_case}
\alias{to_lower_camel_case}
\alias{to_upper_camel_case}
\alias{to_pascal_case}
\alias{to_snake_case}
\alias{to_snek_case}
\alias{to_title_case}
\alias{to_train_case}
\alias{to_shouty_kebab_case}
\alias{to_shouty_snake_case}
\alias{caseconverter}
\title{Convert string depending case preference}
\usage{
to_kebab_case(x)

to_lower_camel_case(x)

to_upper_camel_case(x)

to_pascal_case(x)

to_snake_case(x)

to_snek_case(x)

to_title_case(x)

to_train_case(x)

to_shouty_kebab_case(x)

to_shouty_snake_case(x)
}
\arguments{
\item{x}{A character vector to be converted.}
}
\value{
\itemize{
\item \code{to_kebab_case()} returns character vector in kebab case i.e. "kebab-case"
\item \code{to_lower_camel_case()} returns character vector in lower camel case i.e. "lowerCamelCase"
\item \code{to_upper_camel_case()} returns character vector in upper camel case i.e. "UpperCammelCase"
\item \code{to_pascal_case()} returns character vector in pascal case i.e. "PascalCase"
\item \code{to_snake_case()} returns character vector in snake case i.e. "snake_case"
\item \code{to_snek_case()} returns character vector in snake case i.e. "snek_case"
\item \code{to_title_case()} returns character vector in title case i.e. "Title Case"
\item \code{to_train_case()} returns character vector in train case i.e. "Train-Case"
\item \code{to_shouty_kebab_case()} returns character vector in shouty kebab case i.e. "SHOUTY-KEBAB-CASE"
\item \code{to_shouty_snake_case()} returns character vector in shouty snake case i.e. "SHOUTY_SNAKE_CASE"
}
}
\description{
Convert string depending case preference
}
\examples{
x <- "Demo String"

to_kebab_case(x)

to_lower_camel_case(x)
to_upper_camel_case(x)

to_pascal_case(x)

to_snake_case(x)
to_snek_case(x)

to_title_case(x)
to_train_case(x)

to_shouty_kebab_case(x)
to_shouty_snake_case(x)

}
