% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{heck-package}
\alias{heck}
\alias{heck-package}
\title{heck: Thin Wrapper around 'rust crate heck'}
\description{
Convert strings into cases like snake or camel using 'rust crate heck' \url{https://github.com/withoutboats/heck}.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

}
