% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{heck-package}
\alias{heck}
\alias{heck-package}
\title{heck: Thin Wrapper around `rust crate heck`}
\description{
Convert string into cases like snake using `rust crate heck` \url{https://crates.io/crates/heck}.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
