% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{heck-package}
\alias{heck}
\alias{heck-package}
\title{heck: Highly Performant String Case Converter}
\description{
Provides a case conversion between common cases like CamelCase and snake_case. Using the 'rust crate heck' \url{https://github.com/withoutboats/heck} as the backend to provide highly performant case conversion for 'R'.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

}
